Moved loop checker
When you send a request to the CSE, it will give you back a certain
number of results.  Divide this by ten, and you will get the number of
pages that you can comb through.

When you specify the number of pages max that you want to comb through,
if this is larger than the possible pages that have been calculated,
the for-loop (for piling the results into a list) will now cut off
after the last page of results have been processed.